Added tests for str and get_piece from piece
