+++ pieces.py
@@ -10,3 +10,10 @@ class Piece:
         
     def get_color(self):
         return self.__color__
+
+    # CHECKS IF TO_ROW AND TO_COL ARE IN POSSIBEL POSITIONS:
+
+    def is_row_col_in_valid_positions(self,to_row,to_col,possible_positions):
+        if (to_row,to_col) in possible_positions:
+            return True
+        else: False
