+++ rook.py
@@ -9,13 +9,6 @@ class Rook(Piece):
     def valid_positions(self,from_row,from_col):
         possible_positions=(self.possible_positions_vd(from_row, from_col)+self.possible_positions_va(from_row, from_col)+self.possible_positions_hr(from_row, from_col)+self.possible_positions_hl(from_row, from_col))
         return possible_positions
-    
-    # CHECKS IF TO_ROW AND TO_COL ARE IN POSSIBEL POSITIONS:
-
-    def is_row_col_in_valid_positions(self,to_row,to_col,possible_positions):
-        if (to_row,to_col) in possible_positions:
-            return True
-        else: False
         
     # POSSIBLE VERTICAL DESCENDANT POSITIONS TO MOVE A ROOK TO:
