+++ tests/test_rook.py
@@ -134,24 +134,6 @@ class TestRook(unittest.TestCase):
         possible=rook.valid_positions(4,4)
         self.assertEqual(possible,[(5,4),(6,4),(7,4),(3,4),(2,4),(1,4),(0,4),(4,5),(4,6),(4,7),(4,3),(4,2),(4,1),(4,0)])
 
-    # IS TO_ROW AND TO_COL IN VALID POSITIONS:
-
-    def test_is_row_col_in_valid_postions_true_path(self):
-        board=Board(for_test=True)
-        rook=Rook('BLACK',board)
-        board.set_piece(4,4,rook)
-        possible_positions=rook.valid_positions(4,4)
-        validation=rook.is_row_col_in_valid_positions(3,4,possible_positions)
-        self.assertTrue(validation)
-
-    def test_is_row_col_in_valid_postions_false_path(self):
-        board=Board(for_test=True)
-        rook=Rook('BLACK',board)
-        board.set_piece(4,4,rook)
-        possible_positions=rook.valid_positions(4,4)
-        validation=rook.is_row_col_in_valid_positions(3,3,possible_positions)
-        self.assertFalse(validation)
-
     # DIAGONAL WORNG MOVES:
 
     def test_move_diagonal_desc_right(self):
